pyproject.toml
[project]
name = "amemgym"
version = "0.1.0"
description = "Interactive Memory Benchmarking for Assistants in Long-horizon Conversations"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "backoff>=2.2.1",
    "loguru>=0.7.3",
    "matplotlib>=3.10.8",
    "openai>=1.102.0",
    "pandas>=2.3.2",
    "pyarrow>=21.0.0",
    "python-dotenv>=1.1.1",
    "seaborn>=0.13.2",
    "tiktoken>=0.12.0",
]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true
